Leaderboard api

+++ app/models/leaderboard.rb
@@ -6,15 +6,19 @@ class Leaderboard < ApplicationRecord
 
   def self.update_leaderboard(room_id:, game_name:, time_seconds:)
     all_entries = Leaderboard.all.where(room_id: room_id)
-    
-    max_time_seconds = all_entries.maximum(:time_seconds)
-    if time_seconds > max_time_seconds
-      message = "Sorry, you didn't make the leaderboard."
-    else
-      entry_to_remove = all_entries.find_by(time_seconds: max_time_seconds)
-      entry_to_remove.destroy
+    if all_entries.count < 10
       Leaderboard.create(room_id: room_id, game_name: game_name, time_seconds: time_seconds)
       message = "Congratulations! You've claimed a spot on the leaderboard!"
+    else
+      max_time_seconds = all_entries.maximum(:time_seconds)
+      if time_seconds > max_time_seconds
+        message = "Sorry, you didn't make the leaderboard."
+      else
+        entry_to_remove = all_entries.find_by(time_seconds: max_time_seconds)
+        entry_to_remove.destroy
+        Leaderboard.create(room_id: room_id, game_name: game_name, time_seconds: time_seconds)
+        message = "Congratulations! You've claimed a spot on the leaderboard!"
+      end
     end
     message 
   end
